refactor(layout): remove unused session handling and update user-avatar import
